requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
greenlet==1.1.2
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.35
tomli==2.0.1
urllib3==1.26.9
